strings to ints

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@
 
 int _atoi(char *s)
 {
-	long int i, n, sign;
+	signed int i, n, sign;
 
 	i = 0;
 	while (s[i] == ' ')
